'Footer component and extracted consts from types
