+++ src/components/App/App.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import ErrorBoundary from "../../hoc/ErrorBoundary/ErrorBoundary";
+import { Footer } from "../Footer/Footer";
 import { Hero } from "../Hero/Hero";
 import { Results } from "../Results/Results";
 import "./App.scss";
@@ -10,6 +11,7 @@ export const App = () => {
       <ErrorBoundary>
         <Hero />
         <Results />
+        <Footer />
       </ErrorBoundary>
     </div>
   );
