+++ src/components/Footer/Footer.scss
@@ -1,0 +1,9 @@
+@use "../../styles/colors";
+
+.footer {
+  display: flex;
+  justify-content: center;
+  align-content: center;
+  padding: 1rem 0;
+  background-color: colors.$light-grey;
+}
