+++ src/components/Hero/Hero.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Logo } from "../../components/Logo/Logo";
 import { Button } from "../../components/UI/Button/Button";
-import { ButtonType } from "../../components/UI/Button/Button.types";
+import { ButtonType } from "../../components/UI/Button/Button.consts";
 import { Search } from "../Search/Search";
 import "./Hero.scss";
 
@@ -11,13 +11,13 @@ export const Hero = () => {
       <div className="hero__inner">
         <div className="hero__top-banner">
           <Logo />
-          <Button variant={ButtonType.secondary} onClick={() => { }}>
+          <Button variant={ButtonType.secondary} onClick={() => {}}>
             + ADD MOVIE
           </Button>
         </div>
         <section className="search__wrapper">
           <h1 className="search__heading">FIND YOUR MOVIE</h1>
-          <Search setQuery={() => { }} onSearch={() => { }} />
+          <Search setQuery={() => {}} onSearch={() => {}} />
         </section>
       </div>
     </div>
