+++ src/components/Logo/Logo.const.ts
@@ -1,0 +1,5 @@
+export const LogoSize = {
+    small: 'small',
+    medium: 'medium',
+    large: 'large'
+}
