+++ src/components/Logo/Logo.types.ts
@@ -1,9 +1,3 @@
-export const LogoSize = {
-    small: 'small',
-    medium: 'medium',
-    large: 'large'
-}
-
 export interface ILogoProps {
     size?: string
 }
