+++ src/components/UI/Button/Button.consts.ts
@@ -1,0 +1,15 @@
+export const ButtonType = {
+    primary: "primary",
+    secondary: "secondary",
+};
+
+export const ButtonSize = {
+    small: "small",
+    medium: "medium",
+    large: 'large'
+};
+
+export const ButtonShape = {
+    rectangle: "rectangle",
+    round: "round"
+};
